Replaced EntityTagger with EntityTaggerDB impl; replaced EntityExplainer with EntityExplainerDB impl
